iota_verifier_latest

Module runtime_module_metadata

This pass verifies necessary properties for runtime module metadata, i.e. the module metadata used by the IOTA at runtime. A compiled module may contain at most one metadata item, which must be indexed by the IOTA metadata key. If present, the metadata item must be deserializable and must satisfy any additional checks imposed by the runtime metadata version.

Functions

verify_module

Verifies the runtime module metadata of the given module. If the module does not contain any runtime metadata, just pass. If the module contains runtime metadata, it must satisfy the following: