iota_verifier_latest

Module verifier

This module contains the public APIs supported by the bytecode verifier.

Functions

iota_verify_module_metered

Helper for a “canonical” verification of a module.

iota_verify_module_metered_check_timeout_only

Runs the IOTA verifier and checks if the error counts as an IOTA verifier timeout NOTE: this function only check if the verifier error is a timeout All other errors are ignored

iota_verify_module_unmetered